Move test contracts out
