+++ contracts-test/ERC20Approver.sol
@@ -1,7 +1,7 @@
 pragma solidity ^0.5.4;
-import "../wallet/BaseWallet.sol";
-import "../modules/common/BaseModule.sol";
-import "../modules/common/OnlyOwnerModule.sol";
+import "../contracts/wallet/BaseWallet.sol";
+import "../contracts/modules/common/BaseModule.sol";
+import "../contracts/modules/common/OnlyOwnerModule.sol";
 
 contract ERC20Approver is BaseModule, OnlyOwnerModule {
