+++ contracts-test/FakeWallet.sol
@@ -14,7 +14,7 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 pragma solidity ^0.5.4;
-import "../wallet/BaseWallet.sol";
+import "../contracts/wallet/BaseWallet.sol";
 
 /**
  * @title FakeWallet
