+++ contracts-test/TestModuleRelayer.sol
@@ -1,16 +1,16 @@
 pragma solidity ^0.5.4;
-import "../wallet/BaseWallet.sol";
-import "../modules/common/BaseModule.sol";
-import "../modules/common/RelayerModule.sol";
+import "../contracts/wallet/BaseWallet.sol";
+import "../contracts/modules/common/BaseModule.sol";
+import "../contracts/modules/common/RelayerModule.sol";
 
 /**
- * @title TestModule
+ * @title TestModuleRelayer
  * @dev Basic test module subclassing RelayerModule
  * @author Olivier VDB - <olivier@argent.xyz>
  */
-contract TestModule is BaseModule, RelayerModule {
+contract TestModuleRelayer is BaseModule, RelayerModule {
 
-    bytes32 constant NAME = "TestModule";
+    bytes32 constant NAME = "TestModuleRelayer";
 
     bool boolVal;
     uint uintVal;
