+++ contracts-test/TestOnlyOwnerModule.sol
@@ -1,10 +1,10 @@
 pragma solidity ^0.5.4;
-import "../wallet/BaseWallet.sol";
-import "../modules/common/OnlyOwnerModule.sol";
+import "../contracts/wallet/BaseWallet.sol";
+import "../contracts/modules/common/OnlyOwnerModule.sol";
 
 /**
- * @title TestModule
- * @dev Basic test module.
+ * @title TestOnlyOwnerModule
+ * @dev Basic test onlyowner module.
  * @author Julien Niset - <julien@argent.im>
  */
 contract TestOnlyOwnerModule is OnlyOwnerModule {
