+++ contracts-test/TokenConsumer.sol
