+++ package.json
@@ -8,11 +8,12 @@
   "scripts": {
     "compile:lib": "npx etherlime compile --workingDirectory lib --runs=200",
     "compile": "npx etherlime compile --runs=999",
+    "compile:test": "npx etherlime compile --workingDirectory contracts-test --runs=999",
     "ganache": "npx etherlime ganache --gasLimit=10700000 -e 10000",
     "kovan-fork": "npx ganache-cli -i 42 -l 10700000 -f https://kovan.infura.io/v3/$(cat .env | sed -En 's/INFURA_KEY=''\"''([^''\"'']+)''\"''/\\1/p')@16988040",
     "kovan-fork-latest": "npx ganache-cli -i 42 -l 10700000 -f https://kovan.infura.io/v3/$(cat .env | sed -En 's/INFURA_KEY=''\"''([^''\"'']+)''\"''/\\1/p')",
     "test": "npx etherlime test --skip-compilation",
-    "ctest": "npm run compile && npm run test",
+    "ctest": "npm run compile && npm run compile:test && npm run test",
     "provision:lib:artefacts": "bash ./scripts/provision_lib_artefacts.sh",
     "test:coverage": "bash ./scripts/provision_lib_artefacts.sh & COVERAGE=1 npx etherlime coverage --solcVersion 0.5.4 && istanbul check-coverage --statements 71 --branches 61 --functions 73 --lines 71",
     "lint:contracts": "npx ethlint --dir .",
