+++ test/relayer.js
@@ -4,7 +4,7 @@ const { formatBytes32String } = require("ethers").utils;
 const { parseRelayReceipt } = require("../utils/utilities.js");
 
 const Wallet = require("../build/BaseWallet");
-const TestModule = require("../build/TestModule");
+const TestModule = require("../build/TestModuleRelayer");
 const Registry = require("../build/ModuleRegistry");
 const GuardianManager = require("../build/GuardianManager");
 const GuardianStorage = require("../build/GuardianStorage");
